+++ random/src/main/java/net/daporkchop/lib/random/impl/ThreadLocalPRandom.java
@@ -1,7 +1,7 @@
 /*
  * Adapted from The MIT License (MIT)
  *
- * Copyright (c) 2018-2020 DaPorkchop_
+ * Copyright (c) 2018-2022 DaPorkchop_
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation
  * files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy,
@@ -23,7 +23,6 @@ package net.daporkchop.lib.random.impl;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
-import net.daporkchop.lib.common.util.PorkUtil;
 import net.daporkchop.lib.unsafe.PUnsafe;
 
 import java.util.Random;
@@ -197,7 +196,7 @@ public final class ThreadLocalPRandom extends AbstractFastPRandom {
     }
 
     @Override
-    public long nextLong(int origin, int bound) {
+    public long nextLong(long origin, long bound) {
         if (bound <= origin) {
             throw new IllegalArgumentException("max must be greater than min");
         }
